refactor(nexus): identity module for identity shit
