+++ crates/nexus/src/identity/mod.rs
@@ -1,3 +1,6 @@
 pub mod correlation_id;
 pub mod event_id;
 pub mod nexus_id;
+
+pub use correlation_id::CorrelationId;
+pub use event_id::EventRecordId;
