+++ crates/nexus/src/store/record/mod.rs
@@ -1,58 +1,12 @@
 #![allow(dead_code)]
 use serde::{Deserialize, Serialize};
-use std::{default::Default, fmt::Display, ops::Deref, sync::Arc};
-use uuid::Uuid;
+use std::{fmt::Display, ops::Deref, sync::Arc};
 
 pub mod builder;
 pub mod event_metadata;
 pub mod event_record;
 
 pub use event_record::{EventRecord, EventRecordResponse};
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq)]
-pub struct EventRecordId(Uuid);
-
-impl EventRecordId {
-    pub fn as_uuid(&self) -> &Uuid {
-        &self.0
-    }
-
-    pub fn into_inner(self) -> Uuid {
-        self.0
-    }
-}
-
-impl Default for EventRecordId {
-    fn default() -> Self {
-        EventRecordId(Uuid::now_v7())
-    }
-}
-
-impl Display for EventRecordId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl Deref for EventRecordId {
-    type Target = Uuid;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl From<Uuid> for EventRecordId {
-    fn from(id: Uuid) -> Self {
-        Self(id)
-    }
-}
-
-impl From<EventRecordId> for Uuid {
-    fn from(value: EventRecordId) -> Self {
-        value.0
-    }
-}
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq)]
 pub struct StreamId(Arc<String>);
@@ -84,41 +38,6 @@ impl From<String> for StreamId {
 }
 
 impl From<&str> for StreamId {
-    fn from(id: &str) -> Self {
-        Self(Arc::new(id.to_string()))
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq)]
-pub struct CorrelationId(pub Arc<String>);
-
-impl CorrelationId {
-    pub fn new(id: String) -> Self {
-        CorrelationId(Arc::new(id))
-    }
-}
-
-impl Display for CorrelationId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl Deref for CorrelationId {
-    type Target = str;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl From<String> for CorrelationId {
-    fn from(id: String) -> Self {
-        Self(Arc::new(id))
-    }
-}
-
-impl From<&str> for CorrelationId {
     fn from(id: &str) -> Self {
         Self(Arc::new(id.to_string()))
     }
